➖ Eliminate git_snip gem
It is now broken with the latest git gem. The fix is easy, but I'm finding this less useful when using a Squash & Merge strategy, so I'll delete my local branches manually.

I keep the `gsync` function, though, but simplify it to switch to the main branch and collapse the `fetch -p` and `pull` steps into a single `pull -p` step instead.